fix: create shared test utility for React Query tests and update tests to use it
