+++ client/src/features/tournament/components/dashboard/__tests__/DashboardHeader.test.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { render, screen } from '@testing-library/react';
+import { screen } from '@testing-library/react';
+import { render } from '../../../test-utils';
 import DashboardHeader from '../DashboardHeader';
 import { DashboardData } from '../../../types';
