+++ client/src/features/tournament/test-utils.tsx
@@ -1,0 +1,39 @@
+import React, { ReactElement } from 'react';
+import { render, RenderOptions } from '@testing-library/react';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+
+// Create a new QueryClient for each test
+const createTestQueryClient = () => new QueryClient({
+  defaultOptions: {
+    queries: {
+      retry: false,
+      gcTime: 0,
+      staleTime: 0,
+    },
+  },
+});
+
+type CustomRenderOptions = {
+  queryClient?: QueryClient;
+} & Omit<RenderOptions, 'wrapper'>;
+
+function customRender(
+  ui: ReactElement,
+  options: CustomRenderOptions = {}
+) {
+  const { queryClient = createTestQueryClient(), ...renderOptions } = options;
+
+  const Wrapper = ({ children }: { children: React.ReactNode }) => (
+    <QueryClientProvider client={queryClient}>
+      {children}
+    </QueryClientProvider>
+  );
+
+  return render(ui, { wrapper: Wrapper, ...renderOptions });
+}
+
+// Re-export everything from @testing-library/react
+export * from '@testing-library/react';
+
+// Override render method
+export { customRender as render };
